Prepare tests for S3 library
We pass a storeConfig and helper to the tests.

The storeConfig is because for S3 we use local S3 server when testing so
we need to pass in an s3 object to the tests (similar to the fs object
than can be used for the filesystem backing).

We also pass a helper in to the tests. This helper has one method right
now getFile. It is used to get raw objects from the repository. The
current tests assume a filesystem so the helper allows the tests to
fetch from whatever the backend store happens to be.